requirements.txt
google-generativeai 
python-pptx 
Pillow 
requests 
python-dotenv 
groq
streamlit